cluster: propose->commit->apply for topology; tracked Queue.Delete (audit P1, P5)
In cluster mode the protocol handlers used to mutate the local broker FIRST and
replicate afterwards, so a declare/bind/delete that then failed to commit
(election / lost quorum) left the object changed on this node while the client
saw an error - and Queue.Delete confirmed success from an untracked proposal
before it committed, so a deleted durable queue could "resurrect" on the other
replicas after failover.

Now, for replicated (durable, non-exclusive) topology, the handler does NOT
touch the local broker: it validates read-only, proposes a tracked op, waits for
the Raft commit, and apply_op performs the mutation on EVERY node (including
this one). Nothing is changed locally unless the op commits.

- Queue.Declare: read-only precondition (flag mismatch -> 406, illegal DLX ->
  406) then propose + await; Declare-Ok reports the real depth read post-commit.
- Exchange.Declare: propose + await, no local declare first.
- Queue.Bind: durable queue -> read-only existence check of both endpoints,
  then propose + await; non-durable binds locally.
- Queue.Delete: read-only preconditions (404/405/406 incl. if-unused/if-empty),
  then a TRACKED proposal (cluster_replicate_delete_queue now returns a seq)
  and Delete-Ok only after commit (new PENDING_QUEUE_DELETE_OK).
- New broker_exchange_exists() for the read-only bind precondition.

Exclusive/transient queues and standalone keep the local path (correct: they
are node-local and never replicated). Verified: the 3-node cluster-config test
now also asserts a Queue.Delete removes the queue from every node; standalone
declare/bind/delete/lifecycle unchanged; clean under ASan.

+++ include/cluster.h
@@ -532,8 +532,10 @@ uint64_t cluster_replicate_publish_tracked(cluster_node_t *n, const char *vhost,
 
 /* Replicate a durable queue deletion so every node drops it (idempotent apply,
  * carried in the topology snapshot). Returns 0 if accepted. */
-int cluster_replicate_delete_queue(cluster_node_t *n, const char *vhost,
-                                   const char *queue);
+/* Returns a nonzero tracked `seq` (poll with cluster_proposal_status) so the
+ * caller confirms Queue.Delete only after the delete commits, or 0 on failure. */
+uint64_t cluster_replicate_delete_queue(cluster_node_t *n, const char *vhost,
+                                        const char *queue);
 
 /* Replicate a queue's consume watermark so every node drains its replica copies
  * of messages with cluster_id <= watermark. Returns 0 if accepted. */
